Mark classes-with-res.jar as .KATI_RESTAT
classes-with-res.jar uses commit-change-for-toc, which only updates
the output file if it has changed, but the rule is not marked as
restat.  This causes ninja to assume the file was rebuilt for the
current run, but on the next run it sees the old timestamp and thinks
it is dirty.

Mark the rule as restat in ninja with .KATI_RESTAT.

Bug: 79504400
Test: m java && m java
Change-Id: I09be19d05301418850f58eb24094e342109ca3e6