Avoid per-sdk logic in WindowSessionDelegate
The InsetsVisibility class is going away in U, so the per-sdk approach
will not be viable as the S_V2 and T delegates contains InsetsVisibility
args.

Update to use Proxy.newProxyInstance directly which will provide more
flexibility in intercepting method calls based on the name of the method
and not the exact argument types.

This is a combination of
https://github.com/robolectric/robolectric/commit/5ce246db78552bdaefa27f3efb559e813b19bdf5
and
https://github.com/robolectric/robolectric/commit/8537d6c845a8fab1302d051a1194d9fbb18c27d7

PiperOrigin-RevId: 491648117

Test: TGP'd in Google3
Change-Id: Ie0b7856414f89e5ff4cbbc10a9fe0982bd3790d5