Adapt to new signatures for getting font metrics for Paint and create in bitmap shader.
The in development android platform
has modified signatures for several methods relating to getting font metrics for Paint and creating in bitmap shader, which were being shadowed. This causes Robolectric's
annotation validator to fail when inspecting the new android all jar.

This commit is a quick workaround to modify the shadows
to add variants for the new signatures, to placate the annotation validator.

A future change should consider
adding full support for the new `ShadowNativePaint::nGetFontMetrics`, `ShadowNativePaint::nGetFontMetricsInt`, `ShadowNative::nGetFontMetricsInt` and `ShadowNativeBitmapShader::nativeCreate` functionality.

PiperOrigin-RevId: 573044873
Change-Id: I2852e8d19de0484b1d3062f204d7063ab2651f64