Adapt to new signatures for running character advance for Paint.
The in development android platform
has modified signatures for several methods relating to running character advance for Paint, which were being shadowed. This causes Robolectric's
annotation validator to fail when inspecting the new android all jar.

This commit is a quick workaround to modify the shadows
to add variants for the new signatures, to placate the annotation validator.

A future change should:
1. Consider
adding full support for the new `ShadowNativePaint::nGetRunCharacterAdvance` functionality;
2. Make the parameter types of `ShadowNativePaint::nGetRunCharacterAdvance` strong.

Bug: 319192202
Test: SystemUIGoogleRobo(2)RNGTests
Ignore-AOSP-First: This change needs to go first to generate build for
upstream Robolectric jar updates.

Change-Id: If305ae065602159cd7b0fa7912d1eaeea07692a5