Change the ID of `config_headlineFontFamily`.
When a new ID is added in framework's config file (as ag/26708584 did),
the IDs of the configs are subject to change. This CL changes the
hard-coded ID to mitigate the failures of RNG tests. The next step is
that we do need a more stable way to configure those values.

Bug: 332407470
Bug: 332400195
Test: atest -c SystemUIGoogleRobo(2)RNGTests
Change-Id: I656225f42108962a1eb52b26d86ed67730018379