Remove ShadowUserManager.setSecurityErrorOnNextGetAccountsByTypeCall
A new public API,
ShadowUserManager.setSecurityErrorOnNextGetAccountsByTypeCall, was recently
added to Robolectric. However, AccountManager.getAccountsByType only threw an
SecurityException due to a short-lived framework bug in Android M circa 2015.
This bug was subsequently fixed and released in an Android M update.
For all intents and purposes this issue does not exist any more. Remove the
public API to support this failure mode.

Generally for having public shadow APIs to simulate failure modes, these need
to be well-documented, and part of Android framework API contracts.

PiperOrigin-RevId: 638691163
Change-Id: Ibb886ee36dc7f0bffe76b1b41c5db33dd6e074b0